--- src/components/global/cards/cards.riot
@@ -1,21 +1,0 @@
-<cards>
-    <div>
-
-    <div>
-    
-    <script>
-      
-        
-    
-       export default {
-        onBeforeMount( ){
-
-        }
-       }
-    </script> 
-    <style>
-     h1{
-       color:red;
-     }
-    </style>
-</cards>
