+++ src/components/global/filterPokemon/filterPokemon.riot
@@ -15,14 +15,60 @@
     <option value="ice">Ice</option>
     <option value="dragon">Dragon</option>
     </select>
-    <h1> {state.value}</h1>
     
    <script>
     import getPokemons from '../../../service/pokemonService'
-    import getElementsPokemons from '../cards/index'
+    import createPokemonCard from './index'
     import axios from "axios";
+     
+     const div= document.getElementById("myDIV") 
+     let arrayPoke = []
+     let pokeSelect = []
+     
 
-     const pruebaGet= (value) => {
+     const filter= (pTypeSelected, pAbilitySelected) =>{
+        //pokeSelect.shift()
+        console.log(pTypeSelected)
+        let filteredInfomation = []
+        if(!pTypeSelected == ""){
+          console.log("estoy en el primer filtro")
+            filteredInfomation = arrayPoke.filter((p)=> p.data.types[0].type.name === pTypeSelected)
+        }
+        if(!pAbilitySelected == ""){
+            if(filteredInfomation != NaN){
+            filteredInfomation = filteredInfomation.filter((p)=> p.data.abilities[0].ability.name === pAbilitySelected)
+            }else{
+            filteredInfomation = arrayPoke.filter((p)=> p.data.abilities[0].ability.name === pAbilitySelected)
+            }
+        }
+        pokeSelect=filteredInfomation
+       }
+
+      const printPokemon= ()=>{
+              div.innerHTML=''
+              console.log("pokeselect")
+              console.log(pokeSelect)
+              pokeSelect.forEach((p) => { 
+               createPokemonCard(p.data)
+                })
+          }
+        
+        
+    export default {
+      onBeforeMount(state) {
+          this.getPokemonData()
+      },
+
+      onMounted() {
+        const select= document.querySelector('#poke-type')  
+        select.addEventListener('change', function() {
+          const val= this.value;  
+         filter(val, ""),
+         printPokemon()
+        })
+      },
+
+      getPokemonData(value) {
           getPokemons().then(function (response) {
           const data= response.data.results;
             return Promise.all(
@@ -30,31 +76,17 @@
            )
           })
           .then((pokeRes)=>{
-            const pokeList= pokeRes
-            const filter= pokeList.filter((p)=> p.data.types[0].type.name === value)
-            filter.forEach((p) => document.getElementById("container-cards").appendChild( getElementsPokemons(p.data)))
+            arrayPoke = pokeRes
+            arrayPoke.forEach((p) => { 
+               createPokemonCard(p.data)
+                })
+            //console.log(arrayPoke)
           })
          .catch(function (error) {
           // manejar error
           console.log(error);
          })
        }
-
-    export default {
-      onBeforeMount( ) {
-        state: {
-           value: ''
-         }
-      },
-
-      onMounted() {
-        const select= document.querySelector('#poke-type')  
-        select.addEventListener('change', function() {
-          const val= this.value;  
-          console.log('You selected: ', val), pruebaGet(val)
-        })             
-      }
-    }
-   </script> 
-
+    } 
+   </script>
 </filterPokemon>
