+++ src/index.html
@@ -16,8 +16,7 @@
     <main>
       <div is="header" data-riot-component></div>
       <div is="filterPokemon" data-riot-component></div>
-      <section id="container-cards">
-        <div id="myDIV" is="cards" data-riot-component></div>
+      <section id="myDIV" class="container-cards" >
       </section>
     </main>
     <script
