+++ src/style.css
@@ -5,7 +5,7 @@
   min-height: 100vh;
 }
 
-#container-cards{
+.container-cards{
     display: flex;
     flex-direction: row;
     flex-wrap: wrap;
@@ -14,7 +14,7 @@
 }
 .poke-card{
   background-color: transparent;
-  width: 35vh;
+  width: 41vh;
   height: 53vh;
 }
 .card-inner {
